Changed Opcodes to version 1.6.1. Two opcodes changed, two have been deleted and two have been added. UpdateFields updated to 1.6.1, where ten fields where changed, regarding resistances and player stats fields. Changed field62 to holy_res. Some minor fixes.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@43 26674162-90ff-0310-a85b-b64107c8dee9

+++ src/game/Creature.cpp
@@ -470,13 +470,13 @@ void Creature::generateLoot()
 	itemsToGet = rand()%7; // 0 to 6 items
     }
     
-    m_lootMoney = creature_level * (rand()%5 + 1)*sWorld.getRate(RATE_DROP); //generate copper    
+    m_lootMoney = (uint32)(creature_level * (rand()%5 + 1)*sWorld.getRate(RATE_DROP)); //generate copper    
     
     if( itemsToGet == 0 )
 	return; // sorry dude.. nothing for you
 
     // Generate max value
-    MaxLootValue = ((creature_level * (rand()%40+50))/5)*sWorld.getRate(RATE_DROP)+rand()%5+5;
+    MaxLootValue = (int)(((creature_level * (rand()%40+50))/5)*sWorld.getRate(RATE_DROP)+rand()%5+5);
 
     /*
       We need an algorithm that mimic the randomless given the probabilities of each items.
